// src/api/openapi/model/chatMessage.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserEssential } from './userEssential';


export interface ChatMessage { 
    text: any | null;
    id: any | null;
    timestamp: any | null;
    sender: UserEssential;
    recipient?: UserEssential;
    own: any | null;
}