// src/api/openapi/models/chatRecipient.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ChatRecipient { 
    cost: number;
    email: string;
    firstname: string;
    fullname: string;
    id: number;
    innovaphone_pass: string;
    innovaphone_user: string;
    notifications: boolean;
    secondname: string;
    tel: string;
}